refactor(cilock): rename internal/cmd → cli to allow external composition
The cobra command tree at cilock/internal/cmd was unreachable from
outside the cilock module because of Go's internal-package visibility
rule. Renaming to cilock/cli exposes it so that external code (the
rookery-builder's generated main, plus any third-party composer) can
invoke the full cilock CLI surface via:

  attestation.RegisterLegacyAliases()
  cli.Execute()

Same five lines cilock's own main.go uses today — now usable from any
caller.

Mechanical change: git mv + 'package cmd' → 'package cli' across 15
files, single import update in cilock/cmd/cilock/main.go. Test files
(adversarial_test, subjects_test, verify_test) move with the package
since they're intra-package tests. Verified the renamed cilock still
builds and exposes the full subcommand tree (run, verify, sign, etc).
